+++ HospitalMngmnt/src/main/java/com/HMSApp/HospitalMngmnt/entity/Patient.java
@@ -38,7 +38,7 @@ public class Patient {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "patientid")
-	private Long patientid;
+	private Integer patientid;
 
 	@Column(name = "patientname")
 	private String patientname;
