+++ HospitalMngmnt/src/main/java/com/HMSApp/HospitalMngmnt/repository/PatientRepository.java
@@ -3,9 +3,8 @@ package com.HMSApp.HospitalMngmnt.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.HMSApp.HospitalMngmnt.entity.Patient;
-import java.util.List;
 
-public interface PatientRepository extends JpaRepository<Patient, Long> {
+public interface PatientRepository extends JpaRepository<Patient, Integer> {
     public Patient findByEmail(String email);
 
     public Patient findByPatientid(Long patientid);
